server/python/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=14.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
aiomysql>=0.2.0
pymysql>=1.1.0
greenlet>=3.0.0
